requirements.txt
aiogram>=3.10,<4
httpx>=0.27
pyyaml>=6.0
chromadb>=0.5
sentence-transformers>=3.0
fastapi>=0.115.0
uvicorn>=0.30.0
requests>=2.31.0
faster-whisper>=1.0.0
